requirements.txt
boto3
PyMySQL==0.10.0
PyYAML==5.3.1
pandas